+++ plugin-excel/src/test/java/nl/xillio/xill/plugins/excel/datastructures/XillSheetTest.java
@@ -215,4 +215,21 @@ public class XillSheetTest {
         verify(cellRef, times(6)).getColumn();
     }
 
+    /**
+     * Test {@link XillSheet#forceFormulaRecalculation()}
+     */
+    @Test
+    public void testForceFormulaRecalculation(){
+        // Mock
+        Sheet sheet = mock(Sheet.class);
+        XillWorkbook xillWorkbook = mock(XillWorkbook.class);
+        XillSheet xillSheet = new XillSheet(sheet, false, xillWorkbook);
+
+        // Run
+        xillSheet.forceFormulaRecalculation();
+
+        // Verify
+        verify(sheet).setForceFormulaRecalculation(true);
+    }
+
 }