+++ xill-processor/src/main/java/nl/xillio/xill/plugins/string/services/string/RegexService.java
@@ -32,10 +32,9 @@ public interface RegexService {
      * @param value   the haystack
      * @param timeout in milliseconds
      * @return the matcher
-     * @throws IllegalArgumentException    Is thrown when an illegal argument is passed to the regular expression.
      * @throws PatternSyntaxException      Is thrown when the regular expression has a syntax error.
      */
-    Matcher getMatcher(String regex, String value, int timeout) throws IllegalArgumentException;
+    Matcher getMatcher(String regex, String value, int timeout);
 
 
     /**