+++ xill-processor/src/main/java/nl/xillio/xill/plugins/string/services/string/StringUtilityServiceImpl.java
@@ -19,11 +19,6 @@ import com.google.inject.Singleton;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.text.WordUtils;
 
-import java.math.BigInteger;
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.IllegalFormatException;
 import java.util.List;
 
 /**
@@ -31,75 +26,6 @@ import java.util.List;
  */
 @Singleton
 public class StringUtilityServiceImpl implements StringUtilityService {
-
-    @Override
-    public int damlev(final String source, final String target) {
-        int[] workspace = new int[1024];
-        int lenS = source.length();
-        int lenT = target.length();
-
-        if (lenS * lenT > workspace.length) {
-            workspace = new int[(source.length() + 1) * (target.length() + 1)];
-        }
-
-        int lenS1 = lenS + 1;
-        int lenT1 = lenT + 1;
-
-        if (lenT1 == 1) {
-            return lenS1 - 1;
-        }
-        if (lenS1 == 1) {
-            return lenT1 - 1;
-        }
-
-        int[] dl = workspace;
-        int dlIndex = 0;
-        int sPrevIndex = 0, tPrevIndex = 0, rowBefore, min, cost, tmp;
-        int tri = lenS1 + 2;
-
-        // start row with constant
-        for (tmp = 0; tmp < lenT1; tmp++) {
-            dl[dlIndex] = tmp;
-            dlIndex += lenS1;
-        }
-        for (int sIndex = 0; sIndex < lenS; sIndex++) {
-            dlIndex = sIndex + 1;
-            dl[dlIndex] = dlIndex; // start column with constant
-            for (int tIndex = 0; tIndex < lenT; tIndex++) {
-                rowBefore = dlIndex;
-                dlIndex += lenS1;
-                // deletion
-                min = dl[rowBefore] + 1;
-                // insertion
-                tmp = dl[dlIndex - 1] + 1;
-                if (tmp < min) {
-                    min = tmp;
-                }
-                cost = 1;
-                if (source.charAt(sIndex) == target.charAt(tIndex)) {
-                    cost = 0;
-                }
-                if (sIndex > 0 && tIndex > 0) {
-                    if (source.charAt(sIndex) == target.charAt(tPrevIndex) && source.charAt(sPrevIndex) == target.charAt(tIndex)) {
-                        tmp = dl[rowBefore - tri] + cost;
-                        // transposition
-                        if (tmp < min) {
-                            min = tmp;
-                        }
-                    }
-                }
-                // substitution
-                tmp = dl[rowBefore - 1] + cost;
-                if (tmp < min) {
-                    min = tmp;
-                }
-                dl[dlIndex] = min;
-                tPrevIndex = tIndex;
-            }
-            sPrevIndex = sIndex;
-        }
-        return dl[dlIndex];
-    }
 
     @Override
     public boolean contains(final String parent, final String child) {
@@ -112,7 +38,7 @@ public class StringUtilityServiceImpl implements StringUtilityService {
     }
 
     @Override
-    public String format(final String text, final List<Object> args) throws IllegalFormatException {
+    public String format(final String text, final List<Object> args) {
         return String.format(text, args.toArray());
     }