+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [3.5.0] - unreleased
 
 ### Add
+* Create template plugin [CTC-1799]
 * `System.parseJSON()` now parses LIST [CTC-1769]
 * Warn users about invalid asset names [CTC-1798]
 * Support external logging by adding logging handler that exposes all logging events and exceptions [XSVR-50]
@@ -15,8 +16,13 @@ All notable changes to this project will be documented in this file.
 * Correct typing inconsistencies for booleans: now the strings "false" and "0" evaluate to `false` [CTC-1733]
 * Make StringBehavior conform to type conversion specification: a non-empty string will now evaluate to `true`, unless it contains any of the values "false", "0" or "null" [CTC-1865]
 * Deprecate `File.getText()` [CTC-1860]
+* Update File and Stream help files [CTC-1869]
 
 ### Fix
+* Settings file not found for paths with whitespaces [CTC-1887]
+* Upload project throws 'already exists' after 'invalid names' popup [CTC-1871]
+* Name of included robot not shown during debugging [CTC-1885]
+* Multiple changed robots dialogs and robot tab selection issue [CTC-1441]
 * On server threads are not closed when tasks finishes [XSVR-139]
 * Renaming of robot does not work properly on Linux [CTC-1787]
 * `Web.xPath()` result in variable preview is incorrect [CTC-1802]
@@ -29,6 +35,14 @@ All notable changes to this project will be documented in this file.
 * String.absoluteUrl does not allow protocols other than http(s) [CTC-1806]
 * Dialogs do not render on (very) fast machines [CTC-1850]
 * Concurrency pipeline changes dates and ObjectIds to strings [CTC-1833]
+* Parsing String list results from xpath query floods log and is slow [CTC-1885] 
+* Excel.setCellFormula() throws caught error: "Cannot get a numeric value from a text formula cell [CTC-1528]
+* Order of package names in help file is off in Linux [CTC-1937]
+
+## [3.4.2] - 2016-11-04
+
+### Fix
+* Parsing String list results from xpath query floods log and is slow [CTC-1885] 
 
 ## [3.4.1] - 2016-10-28